requirements.txt
tornado==6.0.4
psycopg2-binary==2.8.4
SQLAlchemy==1.3.15
alembic==1.4.2
environs==7.3.0
ipdb==0.13.2
attrs==19.3.0